Simplify Snowstorm
