--- .github/workflows/ci.yaml
@@ -1,98 +1,0 @@
-name: ci
-
-on:
-  push:
-    branches: [ master ]
-  pull_request:
-    branches: [ master ]
-  release:
-    types: [ published ]
-
-jobs:
-  tests:
-    name: tests
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-      - uses: actions/setup-python@v4
-        with:
-          python-version: "3.11"
-      - run: pip install ruff isort black
-      - run: ruff check .
-      - run: black --line-length=120 --check .
-      - run: isort --check --line-length 120 --profile black .
-
-  release_dev:
-    if: github.ref == 'refs/heads/master' && github.event_name == 'push'
-    name: release_dev
-    environment: Development
-    runs-on: ubuntu-latest
-    needs: [tests]
-    steps:
-      - uses: actions/checkout@v3
-        with:
-          fetch-depth: 0
-      - run: echo "container_tag=$(date +'%FT%H%M%SZ')" >> $GITHUB_ENV
-      - uses: docker/login-action@v2
-        with:
-          registry: binkcore.azurecr.io
-          username: binkcore
-          password: ${{ secrets.ACR_BINKCORE_PASSWORD }}
-      - uses: docker/setup-qemu-action@v2
-      - uses: docker/setup-buildx-action@v2
-      - uses: docker/build-push-action@v3
-        with:
-          push: true
-          context: .
-          platforms: linux/amd64, linux/arm64
-          tags: |
-            binkcore.azurecr.io/${{ github.event.repository.name }}:${{ env.container_tag }}
-            binkcore.azurecr.io/${{ github.event.repository.name }}:latest
-
-  release_staging:
-    if: github.event_name == 'release' && github.event.action == 'published'
-    name: release_staging
-    environment: Staging
-    runs-on: ubuntu-latest
-    needs: [tests]
-    steps:
-      - uses: actions/checkout@v3
-        with:
-          fetch-depth: 0
-      - uses: docker/login-action@v2
-        with:
-          registry: binkcore.azurecr.io
-          username: binkcore
-          password: ${{ secrets.ACR_BINKCORE_PASSWORD }}
-      - uses: docker/setup-qemu-action@v2
-      - uses: docker/setup-buildx-action@v2
-      - uses: docker/build-push-action@v3
-        with:
-          push: true
-          context: .
-          platforms: linux/amd64, linux/arm64
-          tags: |
-            binkcore.azurecr.io/${{ github.event.repository.name }}:${{ github.event.release.tag_name }}
-            binkcore.azurecr.io/${{ github.event.repository.name }}:staging-${{ github.event.release.tag_name }}
-            binkcore.azurecr.io/${{ github.event.repository.name }}:staging
-
-  release_production:
-    if: github.event_name == 'release' && github.event.action == 'published'
-    name: release_production
-    environment: Production
-    needs: [release_staging]
-    runs-on: ubuntu-latest
-    container:
-      image: ghcr.io/binkhq/tagbot:latest
-    steps:
-      - run: |-
-          tagbot \
-            --username binkcore \
-            --password ${{ secrets.ACR_BINKCORE_PASSWORD }} \
-            --source binkcore.azurecr.io/${{ github.event.repository.name }}:${{ github.event.release.tag_name }} \
-            --tag prod-${{ github.event.release.tag_name }}
-          tagbot \
-            --username binkcore \
-            --password ${{ secrets.ACR_BINKCORE_PASSWORD }} \
-            --source binkcore.azurecr.io/${{ github.event.repository.name }}:${{ github.event.release.tag_name }} \
-            --tag prod
