+++ .github/workflows/publish.yaml
@@ -1,0 +1,25 @@
+name: publish
+
+on:
+  push:
+    tags: ["v[0-9]+.[0-9]+.[0-9]+"]
+
+jobs:
+  build:
+    uses: binkhq/actions/.github/workflows/python_build.yaml@v1
+    with:
+      python_version: "3.12"
+      application_name: ${{ github.event.repository.name }}
+      application_version: ${{ github.ref_name }}
+    secrets: inherit
+  release:
+    needs: ["build"]
+    uses: binkhq/actions/.github/workflows/python_release.yaml@v1
+    with:
+      source: ${{ github.event.repository.name }}:${{ github.ref_name }}
+      tags: ${{ matrix.environment }}-${{ github.ref_name }},${{ matrix.environment }}
+      environment:  ${{ matrix.environment }}
+    secrets: inherit
+    strategy:
+      matrix:
+        environment: [staging, production]
