--- lbg-stats-frontend/src/app.css
@@ -1,71 +1,0 @@
-:root {
-  font-family: 'Roboto', sans-serif;
-  line-height: 1.2;
-  font-weight: 400;
-  color: rgba(255, 255, 255, 0.87);
-  background: linear-gradient(to right, #141F33, #12425F);
-  font-synthesis: none;
-  text-rendering: optimizeLegibility;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-body {
-  font-size: 18px;
-}
-
-h1 {
-color: white;
-  font-size: 0.8rem;
-  font-weight: 400;
-}
-
-h2 {
-  color: #71F9D3;
-  font-weight: 300;
-  font-size: 1.3rem;
-}
-
-table,
-th,
-td,
-tr,
-thead {
-  text-align: left;
-  align-self: center;
-  font-weight: 300;
-}
-
-table {
-  background-color: rgba(255, 255, 255, 0.05);
-  width: 90%;
-  border: 2px solid #72C8BD;
-  padding: 12px;
-  border-radius: 10px;
-  margin: 0 auto;
-  text-align: center;
-}
-
-th {
-  color: #72C8BD;
-  font-size: 1.2rem;
-}
-
-td {
-  padding: 6px;
-}
-
-select {
-  background-color: transparent;
-  font-family: 'Roboto', sans-serif;
-  color: white;
-  border: #71F9D3 1px solid;
-  border-radius: 10px;
-  margin: 5px;
-  padding: 6px;
-  font-size: 0.8rem;
-  font-weight: 500;
-  text-transform: uppercase;
-  text-align: center;
-  cursor: pointer;
-}
