--- lbg-stats-frontend/src/main.js
@@ -1,7 +1,0 @@
-import './app.css';
-import App from './App.svelte';
-const app = new App({
-    target: document.getElementById('app'),
-});
-export default app;
-//# sourceMappingURL=main.js.map
