--- run_once/README.md
@@ -1,1 +1,0 @@
-Random scripts that helped with something once upon a time, might be useful so rather than discarding. Saving here.
