+++ snowstorm/tasks/__init__.py
@@ -1,0 +1,1 @@
+"""Tasks for Snowstorm."""
