+++ snowstorm/tasks/database_cleanup.py
@@ -1,3 +1,5 @@
+"""Database Cleanup Task."""
+
 import pendulum
 from loguru import logger
 from sqlalchemy.orm import Session
@@ -5,11 +7,15 @@ from sqlalchemy.orm import Session
 from snowstorm.database import APIStats, Events, FreshService, engine
 
 
-class Job_DatabaseCleanup:
+class DatabaseCleanup:
+    """Database Cleanup Task."""
+
     def __init__(self, days: int) -> None:
+        """Initialize DatabaseCleanup."""
         self.days = days
 
     def cleanup(self) -> None:
+        """Remove old records from the database."""
         delta = pendulum.today().subtract(days=self.days)
         with Session(engine) as session:
             apistats = session.query(APIStats).filter(APIStats.date_time <= delta)
@@ -18,7 +24,7 @@ class Job_DatabaseCleanup:
 
             freshservice = session.query(FreshService).filter(FreshService.updated_at <= delta)
             logger.warning("freshservice records found", extra={"record_count": freshservice.count()})
-            freshservice.delete
+            freshservice.delete()
 
             events = session.query(Events).filter(Events.event_date_time <= delta)
             logger.warning("events records found", extra={"record_count": events.count()})
